refactor(tokenizers): move roberta tokenizer trainer logic to gpt2 tokenizer trainer
