+++ formerbox/data/tokenizers/tokenization_roberta_trainer.py
@@ -1,104 +1,17 @@
 import logging
-from dataclasses import dataclass, field
 from pathlib import Path
-from typing import Any, Dict, List, Optional, Text, Union
+from typing import Any, Optional, Text, Union
 
-from formerbox.common.dataclass_argparse import DataclassBase
+from formerbox.data.tokenizers.tokenization_gpt2_trainer import GPT2TokenizerTrainer
 from formerbox.data.tokenizers.tokenization_roberta import RobertaTokenizer
-from formerbox.data.tokenizers.tokenization_trainer import TokenizerTrainerBase
 from formerbox.modules import TokenizerTrainer
-
-from tokenizers.implementations import ByteLevelBPETokenizer
 
 logger = logging.getLogger(__name__)
 
-
+# pylint: disable=arguments-differ
 @TokenizerTrainer.register(name="roberta", constructor="from_partial")
-class RobertaTokenizerTrainer(TokenizerTrainerBase):
-    # pylint: disable=arguments-differ
-    @dataclass
-    class Params(DataclassBase):
-        files: Optional[List[Text]] = field(
-            default=None,
-            metadata={"help": "The input text files to train a tokenizer on."},
-        )
-        vocab_size: Optional[int] = field(
-            default=None,
-            metadata={"help": "The size of a trained tokenizer's vocabulary."},
-        )
-        min_frequency: int = field(
-            default=2,
-            metadata={"help": "The min frequency for calculating subwords merges."},
-        )
-        legacy_format: bool = field(
-            default=True,
-            metadata={
-                "help": "Whether to save the tokenizer in legacy format (default),"
-                " i.e. with tokenizer specific vocabulary and separate added_tokens files"
-                " or in the unified JSON file format of the `tokenizers` library."
-            },
-        )
-        save_directory: Optional[Text] = field(
-            default=None,
-            metadata={"help": "A path for saving the pre-trained tokenizer."},
-        )
-        add_prefix_space: bool = field(
-            default=False,
-            metadata={"help": "Whether to add a leading space to the first word."},
-        )
-        trim_offsets: bool = field(
-            default=True,
-            metadata={
-                "help": (
-                    "Whether the post processing step should trim"
-                    " offsets to avoid including whitespaces."
-                )
-            },
-        )
-        lowercase: bool = field(
-            default=False,
-            metadata={"help": "Whether or not to preprocess text in lowercase."},
-        )
-        dropout: Optional[float] = field(
-            default=None,
-            metadata={
-                "help": "The likelihood of dropping a subword during calculating the frequency."
-            },
-        )
-        unicode_normalizer: Optional[Text] = field(
-            default=None,
-            metadata={"help": "The unicode text normalizer. Default is set to `None`."},
-        )
-        continuing_subword_prefix: Optional[Text] = field(
-            default=None,
-            metadata={
-                "help": "The subword prefix used for decoding the words. Default is set to `None`."
-            },
-        )
-        end_of_word_suffix: Optional[Text] = field(
-            default=None,
-            metadata={
-                "help": "The suffix that comes after each word. Default is set to `None`."
-            },
-        )
-
-    params: Params
-    params_type = Params
-
-    def __init__(self, params: Params, **kwargs: Any) -> None:
-        super().__init__(params, **kwargs)
-
-    @classmethod
-    def build_tokenizer(cls, params: Params) -> ByteLevelBPETokenizer:
-        return ByteLevelBPETokenizer(
-            add_prefix_space=params.add_prefix_space,
-            lowercase=params.lowercase,
-            dropout=params.dropout,
-            unicode_normalizer=params.unicode_normalizer,
-            continuing_subword_prefix=params.continuing_subword_prefix,
-            end_of_word_suffix=params.end_of_word_suffix,
-            trim_offsets=params.trim_offsets,
-        )
+class RobertaTokenizerTrainer(GPT2TokenizerTrainer):
+    Params = GPT2TokenizerTrainer.Params
 
     def configure_tokenizer(
         self, tokenizer_path: Union[Text, Path], **kwargs: Any
@@ -126,58 +39,3 @@ class RobertaTokenizerTrainer(TokenizerTrainerBase):
             tokenizer_file=tokenizer_file,
             **kwargs,
         )
-
-    def train_tokenizer(self, *args: Any, **kwargs: Any) -> None:
-        del args, kwargs  # use designated args
-
-        # train a tokenizer model
-        assert self.params.files is not None
-        assert self.params.vocab_size is not None
-        assert isinstance(self.tokenizer, ByteLevelBPETokenizer)
-
-        self.tokenizer.train(
-            files=self.params.files,
-            vocab_size=self.params.vocab_size,
-            min_frequency=self.params.min_frequency,
-            special_tokens=self.special_tokens,
-        )
-
-    def save_pretrained(
-        self, save_directory: Optional[Text] = None, **kwargs: Any
-    ) -> None:
-        # take the directory from params if not specified
-        if save_directory is None:
-            assert self.params.save_directory is not None
-            save_directory = self.params.save_directory
-
-        # get the `legacy_format` argument value
-        legacy_format = kwargs.pop("legacy_format", self.params.legacy_format)
-
-        # save the pretrained tokenizer to `save_directory`
-        super().save_pretrained(
-            save_directory=save_directory,
-            legacy_format=legacy_format,
-            **kwargs,
-        )
-
-    @classmethod
-    def get_tokenizer_args(cls, params: Params) -> Dict[Text, Any]:
-        # prepare a copy of args for the pretrained tokenizer
-        kwargs = vars(params).copy()
-
-        # preserve pretrained tokenizer path
-        kwargs.pop("tokenizer_path", None)
-
-        # remove the legacy pretrained format flag
-        kwargs.pop("legacy_format", None)
-
-        # remove training-stage arguments
-        kwargs.pop("files", None)
-        kwargs.pop("vocab_size", None)
-        kwargs.pop("min_frequency", None)
-        kwargs.pop("dropout", None)
-        kwargs.pop("unicode_normalizer", None)
-        kwargs.pop("continuing_subword_prefix", None)
-        kwargs.pop("end_of_word_suffix", None)
-
-        return kwargs
